+++ Example/ViewController.swift
@@ -23,8 +23,7 @@ class ViewController: UIViewController {
         }
     }
     
-    let gkScene: GKScene = .init()
-    var lastUpdateTime = TimeInterval()
+    let timer = Timer()
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -40,10 +39,6 @@ class ViewController: UIViewController {
             node.humanoid.node(for: .neck)?.eulerAngles = SCNVector3(0, 0, 20 * CGFloat.pi / 180)
             node.humanoid.node(for: .leftShoulder)?.eulerAngles = SCNVector3(0, 0, 40 * CGFloat.pi / 180)
             node.humanoid.node(for: .rightShoulder)?.eulerAngles = SCNVector3(0, 0, 40 * CGFloat.pi / 180)
-            let component = try loader.loadComponent(vrmNode: node)
-            let entity = GKEntity()
-            entity.addComponent(component)
-            gkScene.addEntity(entity)
             
             node.runAction(SCNAction.repeatForever(SCNAction.sequence([
                 SCNAction.rotateBy(x: 0, y: -0.5, z: 0, duration: 0.5),
@@ -82,11 +77,6 @@ class ViewController: UIViewController {
 
 extension ViewController: SCNSceneRendererDelegate {
     func renderer(_ renderer: SCNSceneRenderer, updateAtTime time: TimeInterval) {
-        if lastUpdateTime == 0 {
-            lastUpdateTime = time
-        }
-        let deltaTime: TimeInterval = time - lastUpdateTime
-        lastUpdateTime = time
-        gkScene.entities.forEach({ $0.update(deltaTime: deltaTime) })
+        (renderer.scene as! VRMScene).vrmNode.update(deltaTime: timer.deltaTime(updateAtTime: time))
     }
 }
